Alpha version

+++ app/Http/Controllers/UsersListController.php
@@ -1,10 +1,11 @@
 <?php namespace App\Http\Controllers;
 
+use App\Http\Requests\UsersFormRequest;
 use App\User;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
 
-use App\Http\Requests\UsersFormRequest;
+
 use Bican\Roles\Models\Permission;
 use Bican\Roles\Models\Role;
 
@@ -86,22 +87,31 @@ class UsersListController extends Controller {
 	 */
 	public function create()
 	{
+		if ($this->auth_user->can('add.users')){
 		return view('users.create');
+		}else{
+		return view('errors/403');
+		}
 	}
 
 	/**
 	 * Store a newly created resource in storage.
 	 *
+	 * @param UsersFormRequest $request
 	 * @return Response
 	 */
 	public function store(UsersFormRequest $request)
 	{
-		$input = $request->all();
-		$input['password'] = bcrypt($request->get('password'));
+		if ($this->auth_user->can('add.users')) {
+			$input = $request->all();
+			$input['password'] = bcrypt($request->get('password'));
 
 
-		$new_user = User::create($input);
-		User::find($new_user->id)->attachRole(4);
+			$new_user = User::create($input);
+			User::find($new_user->id)->attachRole(4);
+		}else {
+		}
+
 		return redirect('users');
 	}
 
@@ -166,9 +176,9 @@ class UsersListController extends Controller {
 	 */
 	public function update($id, UsersFormRequest $request)
 	{
-//		if(('' === $id) || (!$this->auth_user->is('admin'))){
-//			$id = $this->current_user_id;
-//		}
+		if(('' === $id) || (!$this->auth_user->is('admin|staff', 'all'))){
+			$id = $this->current_user_id;
+		}
 
 
 		// Will return a ModelNotFoundException if no user with that id
@@ -202,24 +212,36 @@ class UsersListController extends Controller {
 	 */
 	public function destroy($id)
 	{
-		//
-		$roles = Role::find('1');
 
-		if (count($roles->users->toArray()) == 1 && $roles->users[0]->id == $id) {
-
-			$error_message = "You are trying to delete only accout you have in your system as Admin Role.
-						\nYou cannot delete this account!";
+		if(('' === $id) || (!$this->auth_user->is('admin|staff', 'all'))){
+			$id = $this->current_user_id;
+		}
+		if($id == $this->current_user_id){
+			$error_message = 'You cannot delete your own account!';
 			\Session::flash('user_delete_error', $error_message);
+			return redirect('users');
+		}		//
+		if ($this->auth_user->can('edit.users')) {
+			$roles = Role::find('1');
+
+			if (count($roles->users->toArray()) == 1 && $roles->users[0]->id == $id) {
+
+				$error_message = "You are trying to delete only accout you have in your system as Admin Role.
+						\nYou cannot delete this account!";
+				\Session::flash('user_delete_error', $error_message);
+
+			} else {
+				$user = User::find($id);
+
+				$user->delete();
+				$message = 'User ' . $user->name . ' deleted!';
+				\Session::flash('flash_message', $message);
+			}
 
 		}else{
-			$user = User::find($id);
-
-			$user->delete();
-			$message = 'User '.$user->name.' deleted!';
-			\Session::flash('flash_message', $message);
+			$error_message = 'You have no permission!';
+			\Session::flash('user_delete_error', $error_message);
 		}
-
-
 
 		return redirect('users');
 	}
